// src/generated/whackend/models/JoinActionDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Agile Wars
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { ActionDto } from './ActionDto';
import {
    ActionDtoFromJSON,
    ActionDtoFromJSONTyped,
    ActionDtoToJSON,
} from './ActionDto';

/**
 * 
 * @export
 * @interface JoinActionDto
 */
export interface JoinActionDto extends ActionDto {
    /**
     * 
     * @type {string}
     * @memberof JoinActionDto
     */
    nickname: string;
}

/**
 * Check if a given object implements the JoinActionDto interface.
 */
export function instanceOfJoinActionDto(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "nickname" in value;

    return isInstance;
}

export function JoinActionDtoFromJSON(json: any): JoinActionDto {
    return JoinActionDtoFromJSONTyped(json, false);
}

export function JoinActionDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): JoinActionDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        ...ActionDtoFromJSONTyped(json, ignoreDiscriminator),
        'nickname': json['nickname'],
    };
}

export function JoinActionDtoToJSON(value?: JoinActionDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        ...ActionDtoToJSON(value),
        'nickname': value.nickname,
    };
}