// src/generated/whackend/models/WeaponDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Agile Wars
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const WeaponDto = {
    Schere: 'SCHERE',
    Stein: 'STEIN',
    Papier: 'PAPIER'
} as const;
export type WeaponDto = typeof WeaponDto[keyof typeof WeaponDto];


export function WeaponDtoFromJSON(json: any): WeaponDto {
    return WeaponDtoFromJSONTyped(json, false);
}

export function WeaponDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): WeaponDto {
    return json as WeaponDto;
}

export function WeaponDtoToJSON(value?: WeaponDto | null): any {
    return value as any;
}